+++ NEWS.md
@@ -6,6 +6,7 @@
 ### New features
 
 * Added a global turn on/off button for the Filter Panel.
+* Added ability to collapse all filters of an individual dataset.
 
 ### Enhancements